Move options logic to separate file
